Extract large classes from Processors into separate files
